+++ NEWS.md
@@ -13,6 +13,7 @@
 * Fix error in extracting data via formula interface for `t_test()` and
   `cohens_d()`.
 * Fix formula interface for `cohens_d()` for within-subject data.
+* Fix check for `htest` type in `chisq_apa()`, `cor_apa()`, and  `t_apa()`.
 
 # apa 0.3.5